fix(build): clear the 68 Kotlin compile errors from the first successful resource link
- SkeuoPanel is open so HudView and HubPanel can extend it (this also restored their material/cornerRadius/options setters).
- Import FrameLayout where LayoutParams are built, and reach getString through the view's context.
- Paint.ascent() is a method, not a property; Rect.centerX/centerY are Int and need toFloat() for drawCircle.
- drawHairline assigned its own parameter; the colour now goes through a local.
- Import the internal SkeuoPathHolder in the widget package.
- Shizuku API 13: the listeners are OnBinderReceived/OnBinderDeadListener, newProcess is private and Sui ships outside the artifact, so both are reached reflectively and degrade cleanly when absent.
- LaunchOutcome is top level, not nested in RobloxLauncher.
- Expression bodies that return early become block bodies (launchPackage, removeView).
- removeThermalStatusListener takes only the listener; the battery default is BATTERY_STATUS_UNKNOWN.
- actionRow gained the enabled flag ControlsActivity passes; ToggleSwitch is set through setChecked since its setter is private; scrollToEnd captures the view instead of a non-existent it.
- Qualify this inside apply/card blocks where the receiver is the layout, not the activity.

+++ app/src/main/java/com/dynablox/launcher/design/widgets/IconTile.kt
@@ -217,9 +217,9 @@ class IconTile @JvmOverloads constructor(
         paint.reset()
         paint.isAntiAlias = true
         paint.color = Tint.alphaFraction(Color.BLACK, 0.55f)
-        canvas.drawCircle(starGlyph.centerX(), starGlyph.centerY(), size * 0.62f, paint)
+        canvas.drawCircle(starGlyph.centerX().toFloat(), starGlyph.centerY().toFloat(), size * 0.62f, paint)
         paint.color = tokens.metalLight
-        canvas.drawCircle(starGlyph.centerX(), starGlyph.centerY(), size * 0.56f, paint)
+        canvas.drawCircle(starGlyph.centerX().toFloat(), starGlyph.centerY().toFloat(), size * 0.56f, paint)
         star.bounds = starGlyph
         star.setTint(tokens.warning)
         star.draw(canvas)
